pull-resources init, saving/sharing progress.
